Try wrapping child components
