+++ src/web/screens/dashboard.tsx
@@ -2,7 +2,7 @@ import { Outlet, useLocation } from 'react-router-dom';
 
 import { SideNav } from '../components/Core/SideNav';
 import { SnailTrail } from '../components/Core/SnailTrail';
-import { PortalRoute } from './routeTypes';
+import { makePrivateRoute, PortalRoute } from './routeUtils';
 import { TeamMembersRoute } from './teamMembers';
 
 export const DashboardMainRoute: PortalRoute = {
@@ -10,7 +10,9 @@ export const DashboardMainRoute: PortalRoute = {
   description: 'Dashboard',
   element: <span>Main</span>,
 };
-export const DashboardRoutes: PortalRoute[] = [TeamMembersRoute, DashboardMainRoute];
+export const DashboardRoutes: PortalRoute[] = [TeamMembersRoute, DashboardMainRoute].map(
+  makePrivateRoute
+);
 const menu = DashboardRoutes.filter((r) => r.description);
 
 function Dashboard() {
