--- src/web/screens/routeTypes.ts
@@ -1,7 +1,0 @@
-import { RouteObject } from 'react-router-dom';
-
-export type PortalRoute = RouteObject & {
-  path: string;
-  element: JSX.Element;
-  description: string;
-};
